+++ examples/test.tf
@@ -7,7 +7,7 @@ terraform {
   required_providers {
     packetfabric = {
       source  = "PacketFabric/packetfabric"
-      version = ">= 1.0.1"
+      version = ">= 1.0.2"
     }
   }
 }