+++ internal/provider/resource_azure_cloud_router_connection.go
@@ -105,6 +105,15 @@ func resourceAzureExpressRouteConnCreate(ctx context.Context, d *schema.Resource
 		if err != nil {
 			return diag.FromErr(err)
 		}
+		createOkCh := make(chan bool)
+		defer close(createOkCh)
+		fn := func() (*packetfabric.ServiceState, error) {
+			return c.GetCloudConnectionStatus(cid.(string), resp.CloudCircuitID)
+		}
+		go c.CheckServiceStatus(createOkCh, fn)
+		if !<-createOkCh {
+			return diag.FromErr(err)
+		}
 		if resp != nil {
 			d.SetId(resp.CloudCircuitID)
 		}